facturacion90porc

+++ Services/Facturacion/Impuesto.cs
@@ -9,16 +9,16 @@ namespace sistema_de_facturacion.Services.Facturacion
     public class Impuesto
     {
         private string concepto;
-        private double valor;
+        private decimal valor;
         
 
-        public Impuesto(string concepto, double valor)
+        public Impuesto(string concepto, decimal valor)
         {
             this.concepto = concepto;
             this.valor = valor;
         }
 
         public string Concepto { get => concepto; }
-        public double Valor { get => valor; }
+        public decimal Valor { get => valor; }
     }
 }
